test: real-backend fixtures + fix struct/dead-field bugs they surfaced
Add RenderingTestFixture (+ utils/gltfio variants) on Engine.Backend.DEFAULT
to cover the GPU-resource binding paths the NOOP driver panics on. Tests skip
gracefully when no backend is available. New real-backend clusters (green on
Metal): Material build/getters/instances, Texture upload+mipmaps,
RenderableManager, Renderer full-frame + readPixels round-trip, IBLPrefilter
run() via a synthetic texture, and gltfio UbershaderProvider.getMaterial.

Bugs surfaced and fixed:
- Material.getParameters() read garbage: the C struct FilaMaterialParameterInfo
  didn't match upstream Material::ParameterInfo. Reworked the bridge to copy
  field-by-field so our struct is layout-independent of upstream.
- Added static_assert layout guards for the remaining reinterpret_cast'd PODs
  (Bone, MaterialKey) so future upstream drift fails the build, not at runtime.
- ViewOptionsRoundTripTest round-trips every field of each View option struct;
  it flagged AmbientOcclusionOptions.scale (exposed but never marshalled — no
  such upstream field) which is removed across all targets, and minConeAngle is
  renamed to minHorizonAngleRad to match upstream.

Removed the @Ignore'd gltfio test; createMaterialInstance's abort is inside
prebuilt filament (documented in MaterialProviderRenderingTest), not a wrapper
bug. HDR/KTX loader paths still need real bundled assets.

+++ kotlin/filament/src/nativeMain/kotlin/io/github/erkko68/filament/Material.native.kt
@@ -130,7 +130,7 @@ actual class Material constructor(internal var nativeHandle: CPointer<FilaMateri
             val info = infoArray[i]
             Parameter(
                 info.name?.toKString() ?: "",
-                Parameter.Type.values()[info.type],
+                Parameter.Type.values()[info.type.toInt()],
                 Parameter.Precision.values()[info.precision.toInt()],
                 info.count.toInt()
             )
